+++ tests/test_plan_element.py
@@ -1,7 +1,5 @@
 from dataclasses import dataclass
 
-import pytest
-
 import lxcraft
 
 
@@ -42,6 +40,3 @@ def test_plan_element():
 
         plan.execute()
         plan.execute()
-
-    with pytest.raises(Exception, match=r"already has a plan"):
-        lxcraft.Plan(element)